Fixes base problem of #117
Why does this fix the problem? I guess because this function was written
by crazy people. Apparently if it converts a string into XML it doesn't
automatically enforce the required structure. So, by giving it one big
XML container to sit inside. it negates the issue. This is an avowedly
stupid solution, but as far as I've been able to tell, it's the only one
that works, so this one is on you PHP.