Fix remote page retrieval in AB_Subscription_Builder
When switching to wp_remote_get() in 745da9b8e, a switch was also made
away from file_get_html(), a function defined by the simple_html_dom
library. As a result, the object being returned was an HTML string rather
than a DOM document. This was causing fatal errors when running the
Academic Blogs subscription builder script.

The fix is to create a wrapper that uses simple_html_dom, and then to
call this wrapper whenever such a DOM object is required.