Filtering the feed cache time.
For some reason, filtering the feed cache time didn't matter in previous testing, it was retrieving the feed regularly without a problem. However, in further testing, it has. Why this wouldn't have happened before isn't clear. 

The value is set to 45 minutes. If this doesn't cause any problems for users, we should drop it to 30 minutes later.