Allow sites w/normal usr capabilities to be normal
Though the miracle of my reverse role to capability lookup system is
clearly the (somewhat tested) solution for all the sites with the cool
capability settings, I guess we should probably let the normal sites
use normal role setting. It isn’t nearly as exciting though. I guess it
is more stable, but really, that’s not as much fun.