Move btn-warning alterations to override
from our CSS.

Hey @regan008 I'm moving your changes to the btn-warning class to our
style sheet. Since the bootstrap.css file is a library we've brought in
from elsewhere and might like to upgrade or maintain, we don't want to
put any code we write into that file. Instead I'm putting it in the
style.css file where, because it is called later, it will override the
bootstrap.css file.

This will make it easier to maintain and to alter in the future. Sorry I
didn't give you the heads up about this earlier!