Respect the switchable state of schema bits
Resolves #553

What this does: any schema object can have a switchable state (set by
class). If it is switchable it can be toggled on to off, off to on,
over and over. If not, it can only be turned to On.

The `read` schema is switchable (it can be on or off, and turned both
ways) and the Mark As Read button reflects this. But the act of
actually *reading* isn’t switchable, and so opening the modal and
nominating the box shouldn’t switch the read state back and forth.

Because the H1 a element doesn’t actually have its own state, it uses a
param to indicate which element and schemas it should alter, but the
code was pulling switchable-ness from the targeted schema bits, not the
actual acted on item.

With this change the acted on item should be the one that controls the
switchable state of a schema bit. So the Mark As Read item is set to
switchable, and can switch the read state, but the H1 a element isn’t
set to switchable and so its interactions with the schema state aren’t
switchable.

This is a pretty cool thing to do and a pretty cool way to do it (not
to toot my on horn here), and I feel it is decently elegant, which
isn’t often the case with my javascript. It can do a lot with very
little, which is why I’m explaining it all here, because the code
doesn’t fully indicate the opportunity of use.