No duplicate content on infini scroll
This resolves #566

The problem was that in cases where the retrieval process was actively
occurring more items were being added to the database while the user was
on the All Content page. The content would push down the current items
in the query so that they would be duplicated on page in the infinite
scroll retrieval. These duplicate IDs were containing items skipped over
by the script that straightened out the placement of the action buttons
(because duplicate IDs are illegal).

This change removes those duplicate ID-tagged articles after infinite
scroll is complete.