Allow users to refresh feeds from the Edit Posts table list of feeds
Resolves #467 by allowing users to manually refresh individual feeds.

Doing so will only occur in a refresh under the following conditions:

1. A feed chunk is not actively open
2. The feed retrieval process is not in the midst of attempting to
determine the next feed to chunk.
3. The feed being manually refreshed was neither just retrieved or the
previous feed retrieved as part of the retrieval process (in order to
avoid duplicated content caused by the feed being retrieved twice
simultaneously)

Additionally, this commit resolves a bad class call in the retrieval
process.