Post we are trying to Draft already exists? Quietly fail posting and refer the user to the originally created post
This resolves #459 pretty well, though we should measure somewhere if
something has been passed through this process more than once by more
than one person, as that might indicate something useful.

For now, it will fail silently, but respond as if the user drafted the
older version of the post; This is better than earlier fails, which
were nonsense and useless.