Moving all meta handling under the new system and activating serialization
BIG NOTE: This means that all meta properties in the PressForward
system are now handled by our built-in meta processor, which interacts
with WP through an interface. It also means that not all pf meta keys
exist as WordPress Meta Keys, instead some are stored under their own
fields and some are stored, serialized, under pf_meta, the master meta
key.

This makes it essential that pf-meta's be accessed through pf functions
from here on out, some fields will no longer be available by WP's
access functions without modification.

Why are we doing this? The sheer scale of PF metas is incredibly hard
on WP databases, especially for sites with large subscriptions. This is
causing slowness and other difficulties for some users and it is
essential that we decrease our system's footprint in the meta database
table.