This solves the OPML bug by fixing attempt to get all metas.
This resolves #810 but not to the level of completion I'd like. It
still will throw errors (silently if WP_DEBUG is off) if there are no
folders in the system. Those errors should be resolved, but I'm
thinking it is best to wait until next release for these in order to
respect an attempt at schedule.