Give Nominate This its own permission
Ok, so there are so many damn interactions here that could go wrong
depending on how user role structures are being managed and where the
subject site could potentially intercept roles. Arguably this is a good
reason to just make up a whole new set of roles for everything
PressForward but honestly... must we? Instead we're going to create a
new user level option for the Nominate This bookmarklet hinged on the
main permission sets with attention and support for advanced role
management through the existing system. Let's talk about a whole new
permission set at some future point but solve this now.

#893