Fix incorrect `$this` references.
The instance in Feed_Items.php is likely a typo and is probably causing fatal
errors in some cases.

The instances in Nominated.php are due to PHP throwing notices when you use
`$this` in the context of a method called statically. In general, swapping out
for `self::` looks like it will work in all of these cases, but this could use
another look by someone who understands the code a bit better.