Initial pass at client-side Readability parsing in bookmarklet.
Because of CORS restrictions in place on many sites, we use the
WP server as a proxy. Nominate This makes an AJAX request to WP;
WP fetches the page; WP returns the raw request result to the Nominate
This window. Content is then parsed using Readability.js in the client.

Author, tags, and featured image are also parsed in the client.

See #1097.