Attempting to fix the double-loading of reader modals.
Allowing this block to fire for single-item modals resulted in the
creation of a duplicate modal, one of which was positioned over the top
of the other and made it impossible to interact with the visible one.

It's still necessary to allow this callback to fire for comment modals.

See #1151.