On Reader view, only load images when expanding the modal.
We do this by removing the `src` attributes from image tags
when building Reader markup, stashing the URLs in a `data-src` attribute.
Then, when the modal is opened, we swap the `src` attributes back in.

This dramatically improves Reader performance when loading a list
of items that contains many images, especially those that are
loaded from servers with latency issues.

See #1114, #1054.