When building Reader interface, don't attempt to fetch embed HTML for non-embeddable URLs.
A feature was added in a79f70de687a569c532aa8f692756fb8d049d80a that enabled
an oEmbed-powered preview if the nominated/feed URL was an oEmbed-capable URL
(as in, say, a YouTube URL). But the implementation was written in such
a way that *all* submitted item URLs were checked for oEmbed periodically.
Generally, this is not necessary: we can prescreen by checking against
WP's list of registered handlers.

See #1054. See #888.