Rework sorting mechanism for Reader interface.
jquery-tinysort only sorted items that appeared in the DOM, which was
not compatible with pagination.

The client-side sorting has been replaced with new 'sort-by' and 'sort-order'
URL parameters, which are handled server-side on both the All Content
and Nominated screens.

This change means using a new mechanism for sorting by 'Date nominated',
since PF wasn't properly storing nomination date in a way that allowed
for sorting. It is now stored in 'sortable_nom_date', and is only stored
the first time that an item is nominated. Subsequent nominations do not
change this initial value. As such, 'Date nominated' will always look
only at the first nomination date.

See #1031. See #1156.