Fix quick-edit modifications, and make them i18n-friendly.
The quick-edit modifications were broken since WordPress changed the
'Quick Edit' link into a button. This changeset fixes the event listener.

In addition, this changeset moves the quick-edit script into the NPM
build pipeline, so that we can use WP's i18n libraries and properly
internationalize the strings in the script.

See #1117, #1159.