Move delete queue processing to a cron job.
Previously, items in the delete queue were processed using a self-ping.
This setup required the use of a bespoke nonce system. It was also
subject to the vagaries of self-pings, which can be blocked on some
server configurations.

This changeset keeps the delete queue system, but processes it instead
on an hourly 'pf_process_delete_queue' cron task.

See #1113.