Module module_setup() routine for modules.
Previously, each module's `module_setup()` method saved default
options to the database on every save. This is not necessary,
and it results in unneeded database queries.

Instead, the `module_setup()` logic is abstracted in the base
module class to check for the existing of default settings before
saving the defaults. Then, each module has a new `get_default_settings()`
method to provide its default settings.

See #1146.