Some standardization of parameters in Controllers\Metas.
The `$state` parameter on `add_meta()` and related methods
was being used inconsistently, sometimes as a uniqueness boolean
and sometimes as a value check (ie, only update a piece of metadata
if the key AND value match). I've eliminated the latter types of checks
and standardized on the boolean `$is_unique` treatment throughout.

See #1167.