Refined logic for `pressforward_use_block_nominate_this()`.
In order to more closely follow the settings in the Classic Editor
plugin, we mimic the core filters that WP itself uses to determine
which editor to use. Then plugins like Classic Editor filter them,
and we use the filtered values to determine whether to use the
block or classic editor.

This change means that the config setting will become "Force Classic
editor for Nominate This", and the default behavior will be Block,
where Block is available.

See #1124.