Nominated panel should show 'nomination' items from both draft and publish status.
The Classic Nominate This interface places nominations in draft status.
The Block Nominate This interface places them in publish status. It may
turn out to be necessary to regularize this in the future, but for now
the multiple values of 'post_status' when assembling the Notifications
panel allows them all to be visible.

See #1124.