Introduce notification emails for users on new comments.
Users receive notifications on comments that are posted to nominations
or other items of which they are the authors, or on which they've
previously commented.

Users may toggle these notifications on or off in their user settings.
A global fallback setting is available on Dashboard > PressForward > Tools >
Site Options, alongside other global email notification defaults.

See #1044.