Use `post_title` to populate 'text' attribute in OPML.
The use of `post_content` caused problems with inline error messages,
as well as overlong post content, HTML tags, and other content
that is invalid or unnecessary in the context of an OPML.

The `'text'` attribute is used by clients such as RSS readers
to display the title of the feed, and as such, `post_title` is a more
appropriate choice.

See #1217.