Prevent recursion when mapping 'edit_posts' capability.
Since #1223, a mapped capability has been used to grant users with the ability
to add tags to nominations. The primitive cap used here is 'edit_posts', which
WP itself uses as a heuristic for whether a user ought to be able to add
tags to content. However, the change introduced a problem when the custom
Nominate This role was set to Contributor, because Contributors are identified
in the PF codebase by the 'edit_posts' capability. As a result, the logic introduce
in dee173f0f38bbbe944a7001fa2a427cf04b11365 caused an infinite loop.

We fix this by unhooking the cap-mapping filter and checking the stored (unfiltered)
caps for the current user.

See #1234.